Add test for if two countries have the same or different volunteers per event
We might want to see if beach cleanup events tend to get more or fewer
people per event in two countries. This runs a t-test or a Z-test to
find out. More details are given in the different-volunteers notebook,
and you can use the function in the module ml/tests.py.