Predict day of week of cleanup
An organisation might want to pick a specific day of the week for
their cleanup. We create a model which takes in information about
their upcoming cleanup, such as its location, to determine which
day of the week to use.

We implement a ZeroR and OneR model so far, and we can implement
more powerful models soon.